feat: first attempt at upgrading to Gatsby v5
Does not still work locally so committed this under feature branch for the time being.

When building the site All I get is this error:

`"Error: error:0308010C:digital envelope routines::unsupported"`

It's somehow related to Webpack (sigh). Stack Overflow advises using a legacy OpenSSL provider because apparently a security hole was patched in Node v17. That workaround is questionable at best. The fun part is how Gatsby just released v5 and it supports only Node.js v18 so downgrading to v16 is not a solution either.